backend/requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
Flask-CORS==4.0.0

# Database
# psycopg2-binary==2.9.6  # Commented out for local development
SQLAlchemy==2.0.23

# Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Environment
python-dotenv==1.0.0

# Monitoring
psutil==5.9.5

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Code Quality
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0

# Security
cryptography==41.0.4

# Logging
structlog==23.1.0 